refactor: add a utils section for integration tests
