+++ packages/solidity-contracts/integration-tests/transfer_eth.ts
@@ -11,11 +11,11 @@ import type {
 
 // due to cyclic workspace dependencies with test-utils package, the CI which runs `pnpm forc fmt --check` was getting timed out
 // so hence had to re-use some test helpers in the test-utils package
-import { FUEL_CALL_TX_PARAMS } from './constants';
 import { createRelayMessageParams, waitForBlockFinalization } from './ethers';
 import { waitForMessage, getMessageOutReceipt, getBlock } from './fuels';
 import { setupEnvironment, fuels_parseEther } from './setup/setup';
 import type { TestEnvironment } from './setup/setup';
+import { FUEL_CALL_TX_PARAMS } from './utils/constants';
 
 const { expect } = chai;
